rexstan enhancements

+++ lib/category.php
@@ -164,7 +164,7 @@ class Category implements \D2U_Helper\ITranslationHelper
                         .'ON main.category_id = default_lang.category_id AND default_lang.clang_id = '. rex_config::get('d2u_helper', 'default_lang') .' '
                     .'WHERE target_lang.category_id IS NULL '
                     .'ORDER BY default_lang.name';
-            $clang_id = rex_config::get('d2u_helper', 'default_lang');
+            $clang_id = (int) rex_config::get('d2u_helper', 'default_lang');
         }
         $result = rex_sql::factory();
         $result->setQuery($query);
