+++ plugins/news_types/lib/types.php
@@ -164,7 +164,7 @@ class Type implements \D2U_Helper\ITranslationHelper
                         .'ON main.type_id = default_lang.type_id AND default_lang.clang_id = '. rex_config::get('d2u_helper', 'default_lang') .' '
                     .'WHERE target_lang.type_id IS NULL '
                     .'ORDER BY default_lang.name';
-            $clang_id = rex_config::get('d2u_helper', 'default_lang');
+            $clang_id = (int) rex_config::get('d2u_helper', 'default_lang');
         }
         $result = rex_sql::factory();
         $result->setQuery($query);
